Obstacles move from side to side
added instructions in the start method for movement on the x axis according to height. The equation is not perfect and needs to take the obstacles scale into acocunt to work as intended.

Next step is to figure out how to add different bounce behaviour for different shapes. As well as different colliders and sprites depending on shape.