Finished meny, multiplier, start and game over of the game
Continuing work on minimum required product.

Next step is to make obstacles spawn at different altitudes and longitudes.
After that I add the skip and make it work with the multiplier system.
Third is to make circles and triangles bounce.
Forth is to add light and particleeffects to make the game pretty.
Fifth is to add cubes.

I'm on my way!