Made circles and triangles bounce differently
Circles now allways bounce in the same direction and triangles allways bounce in the direction of the player.

Fine tuning is still needed for the bounces to feel predictable and fluid.

Next task is to make triangles touch the floor and to make big obstacles bounce correctly (higher than other shapes