A bit of rought fine tuning
Obstacles now bounce up to a set height determined by their size each bounce.

Triangles bounce to the side even if they frop from straight down.

Lives are set to 1 at the start of the game.

Next is to fix triangles so that they touch the ground the same way that circles do.