Triangle bounce fix and Post processing
Triangles now touch the ground before bouncing.

Added PostProcessing and a bloom effect.

Have still not been able to make pointBalls and multiplierCubes spawn anywhere else than in the middle of the screen.

Next session I will call Sam for the point and multiplier issue and I will try to add particles on obstacle-destruction that fall to the ground and give points when picked up by the player.