Experimented with creating PointBalls
Added a new PointBall prefab, a pointballscript and a "spawnPointBallsScript" to run a method to add many pointballs from a position at the same time.

Next task is to create a OnTriggerEnter method to the PointBall to make it add points on trigger with player. After that is to try it out and see what works and what dose not work.